+++ accounts.csv
@@ -1,0 +1,21 @@
+1985-07-24,John Doe,johndoe,Cashier,true
+1990-04-15,Jane Smith,janesmith,Pharmacist,false
+1975-12-05,Mike Johnson,mikej,Admin,true
+1982-03-10,Emily Clark,emilyc,PharmacyManager,true
+1989-11-02,Robert Lee,robertl,PharmacyTechnician,false
+1995-01-22,Anna White,annaw,Admin,true
+2000-06-30,Ben Adams,benad,Cashier,false
+1980-09-18,Lisa Brown,lisab,Pharmacist,true
+1993-02-14,Tom Green,tomg,PharmacyManager,true
+1987-08-09,James Black,jamesb,Cashier,false
+1998-05-23,Susan Grey,susang,Admin,true
+1970-11-29,Michael Taylor,michaelt,Pharmacist,false
+1996-07-15,Karen Miller,karenm,PharmacyTechnician,true
+1994-03-03,David Wilson,davidw,Admin,false
+1983-12-20,Peter Scott,peters,Pharmacist,true
+2001-10-10,Olivia Evans,oliviae,PharmacyManager,false
+1997-05-12,George Harris,georgeh,Cashier,true
+1981-01-28,Patricia King,patriciak,PharmacyTechnician,false
+1992-08-17,William Lewis,williaml,Pharmacist,true
+1984-06-25,Jessica Turner,jessicat,Admin,true
+1999-04-04,Steven Carter,stevenc,PharmacyManager,false
