+++ customers.csv
@@ -1,0 +1,26 @@
+1975-04-18,Jane Smith
+2002-01-29,John Doe
+1976-08-28,John Adams
+1952-05-06,Anna Lee
+1961-08-25,Sarah Lee
+1989-06-15,Elizabeth Jenkins
+1969-06-01,Michael Clark
+1984-12-13,Emily Clark
+1954-10-10,Emily Brown
+1961-11-15,Yyaya Doe
+1966-05-20,Emily Brown
+1962-02-24,Lincoln Brown
+2001-12-27,Collins Clark
+1977-01-29,Ivy Doe
+1955-07-25,Jane Emmy
+1986-03-21,Emfield Smith
+1990-12-18,Jane Dudu
+1964-09-18,Lilly Gloria
+1991-10-01,Anna Lee
+1992-01-31,Kenya Amber
+1978-02-02,Mael Davis
+1975-02-16,Yee Sato
+1953-03-17,Olivia Benson
+1957-04-18,Ivana Nkunzi
+1992-04-16,Annabel Lamza
+2000-12-06,Jason Phys
