+++ drugs.csv
@@ -1,0 +1,35 @@
+2025-12-15,100,15.50,Paracetamol,true,Paracetamol 500mg
+2024-11-10,200,10.75,Ibuprofen,false,Ibuprofen 200mg
+2023-10-05,150,25.00,Aspirin,true,Aspirin 300mg
+2026-03-22,50,5.99,Vitamin C,false,Vitamin C 1000mg
+2023-06-30,30,100.00,Morphine,true,Morphine 10mg
+2027-01-01,200,30.00,Amoxicillin,false,Amoxicillin 500mg
+2024-09-15,120,12.50,Metformin,true,Metformin 850mg
+2022-08-01,80,40.00,Diazepam,true,Diazepam 5mg
+2024-12-20,100,18.75,Omeprazole,false,Omeprazole 20mg
+2025-07-05,150,23.40,Ciprofloxacin,false,Ciprofloxacin 500mg
+2024-05-10,250,12.30,Hydrochlorothiazide,true,Hydrochlorothiazide 25mg
+2025-11-22,180,22.00,Clopidogrel,false,Clopidogrel 75mg
+2023-04-18,60,8.00,Atorvastatin,true,Atorvastatin 20mg
+2024-10-10,90,14.00,Metoprolol,false,Metoprolol 50mg
+2025-02-15,140,17.60,Losartan,true,Losartan 25mg
+2026-06-25,75,10.10,Warfarin,false,Warfarin 5mg
+2024-03-02,130,11.50,Simvastatin,true,Simvastatin 40mg
+2023-07-20,50,19.80,Levothyroxine,true,Levothyroxine 50mcg
+2027-02-05,200,26.20,Amiodarone,false,Amiodarone 100mg
+2024-12-12,180,21.90,Spironolactone,true,Spironolactone 25mg
+2023-11-15,160,27.40,Prednisone,false,Prednisone 10mg
+2026-09-05,100,24.00,Azithromycin,true,Azithromycin 250mg
+2024-08-01,110,18.00,Furosemide,false,Furosemide 40mg
+2025-05-22,130,20.30,Alprazolam,true,Alprazolam 0.5mg
+2026-02-10,140,22.50,Enalapril,false,Enalapril 5mg
+2023-09-10,170,13.20,Fenofibrate,true,Fenofibrate 145mg
+2024-04-15,200,11.00,Losartan,false,Losartan 50mg
+2027-03-01,90,19.00,Hydrochlorothiazide,true,Hydrochlorothiazide 12.5mg
+2023-05-15,80,16.70,Amiodarone,false,Amiodarone 200mg
+2024-11-10,75,22.00,Duloxetine,true,Duloxetine 30mg
+2025-01-01,100,28.50,Fenofibrate,true,Fenofibrate 160mg
+2026-04-20,200,13.10,Carvedilol,false,Carvedilol 12.5mg
+2023-10-15,120,25.30,Tramadol,true,Tramadol 50mg
+2024-02-28,150,10.00,Hydrocodone,false,Hydrocodone 5mg
+2025-08-10,130,21.20,Hydroxychloroquine,true,Hydroxychloroquine 200mg
