+++ stock.csv
@@ -1,0 +1,8 @@
+10,15.99,Shampoo,10
+50,2.50,Toothbrush,20
+30,3.75,Hand Sanitizer,0
+100,0.99,Face Mask,5
+25,5.49,Deodorant,15
+12,8.00,Body Lotion,0
+60,1.25,Paper Towel,30
+75,4.99,Shaving Cream,0
